chore: rename guest and other minor ux improvements
