+++ bin/client-eth/Cargo.toml
@@ -1,0 +1,24 @@
+[workspace.package]
+[package]
+name = "rsp-client-eth"
+description = ""
+edition = "2021"
+
+[dependencies]
+bincode = "1.3.3"
+
+# workspace
+rsp-client-executor = { path = "../../crates/executor/client" }
+
+# sp1
+sp1-zkvm = "1.0.1"
+
+[patch.crates-io]
+# alloy-eips = { path = "../alloy/crates/eips" }
+# alloy-serde = { path = "../alloy/crates/serde" }
+alloy-eips = { git = "https://github.com/jtguibas/alloy.git", branch = "john/rsp-8e9e6ac" }
+alloy-serde = { git = "https://github.com/jtguibas/alloy.git", branch = "john/rsp-8e9e6ac" }
+tiny-keccak = { git = "https://github.com/sp1-patches/tiny-keccak", branch = "patch-v2.0.2" }
+
+# Using the fork until https://github.com/alloy-rs/trie/pull/27 is merged and released
+alloy-trie = { git = "https://github.com/xJonathanLEI/trie.git", rev = "fb57b642278fd4213f244a2ab217be831a51abc9" }
