+++ bin/client-eth/src/main.rs
@@ -1,0 +1,18 @@
+#![no_main]
+sp1_zkvm::entrypoint!(main);
+
+use rsp_client_executor::{io::ClientExecutorInput, ClientExecutor, EthereumVariant};
+
+pub fn main() {
+    // Read the input.
+    let input = sp1_zkvm::io::read_vec();
+    let input = bincode::deserialize::<ClientExecutorInput>(&input).unwrap();
+
+    // Execute the block.
+    let executor = ClientExecutor;
+    let header = executor.execute::<EthereumVariant>(input).expect("failed to execute client");
+    let block_hash = header.hash_slow();
+
+    // Commit the block hash.
+    sp1_zkvm::io::commit(&block_hash);
+}
