+++ crates/executor/client/Cargo.toml
@@ -1,0 +1,48 @@
+[package]
+name = "rsp-client-executor"
+description = ""
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+eyre.workspace = true
+serde_json.workspace = true
+serde.workspace = true
+tokio.workspace = true
+
+# workspace
+rsp-witness-db.workspace = true
+rsp-primitives.workspace = true
+rsp-mpt.workspace = true
+
+# reth
+reth-ethereum-consensus.workspace = true
+reth-optimism-consensus.workspace = true
+reth-execution-types.workspace = true
+reth-primitives = { workspace = true, features = ["k256"] }
+reth-storage-errors.workspace = true
+reth-trie.workspace = true
+reth-evm.workspace = true
+reth-evm-ethereum.workspace = true
+reth-evm-optimism.workspace = true
+reth-errors.workspace = true
+reth-chainspec.workspace = true
+
+# revm
+revm.workspace = true
+revm-primitives.workspace = true
+
+# alloy
+alloy-primitives.workspace = true
+alloy-rlp.workspace = true
+itertools = "0.13.0"
+url.workspace = true
+futures.workspace = true
+
+[dev-dependencies]
